cloud_run/requirements.txt
google-cloud-translate
Flask
gunicorn==20.1.0
google-cloud-storage
python-dotenv